move user schemas
